requirements.txt
dify_plugin>=0.4.0,<0.7.0
pytz>=2024.1
python-dateutil>=2.9.0
requests>=2.31.0
python-dotenv>=1.0.0
